refactor: extract shared utilities and fix data leaks
Extract htmlToMarkdown and buildOutlineTree into reusable libraries
to eliminate duplication across MCP and API export routes. Fix N+1
query patterns in coaching tools and unscoped relationship query that
leaked data across projects.

- Extract htmlToMarkdown into src/lib/html-to-markdown.ts with merged
  entity decoding and beat-stripping support
- Extract buildOutlineTree into src/lib/outline-tree.ts for consistent
  outline generation
- Fix unscoped relationship query in MCP exportStoryBible (data leak)
- Fix N+1 queries in getConsistencyContext and getStoryBibleCrossReference
- Update export routes to use shared utilities
